Seperate out Graph and Summary
